Fixed lockup when spawned too close Added proper subtitles (except for flee - that whole feature is going to be completely reworked)

+++ src/main/java/com/gargin/cavenoise/entity/custom/DwellerStareGoal.java
@@ -60,7 +60,7 @@ public class DwellerStareGoal extends Goal {
         LivingEntity pendingTarget = this.cavedweller.getTarget();
         Minecraft minecraft = Minecraft.getInstance();
         boolean yawPlayerLookingTowards = false;
-        float fov = (float)(Integer)minecraft.options.fov().get();
+        float fov = (float)minecraft.options.fov().get();
         float yFovMod = 0.65F;
         float fovMod = (35.0F / fov - 1.0F) * 0.4F + 1.0F;
         fov *= fovMod;
@@ -68,10 +68,10 @@ public class DwellerStareGoal extends Goal {
         Vec3 b = this.cavedweller.position();
         Vec2 dist = new Vec2((float)b.x - (float)a.x, (float)b.z - (float)a.z);
         dist = dist.normalized();
-        double newAngle = Math.toDegrees(Math.atan2((double)dist.x, (double)dist.y));
+        double newAngle = Math.toDegrees(Math.atan2(dist.x, dist.y));
         float lookX = (float)pendingTarget.getViewVector(1.0F).x;
         float lookZ = (float)pendingTarget.getViewVector(1.0F).z;
-        double newLookAngle = Math.toDegrees(Math.atan2((double)lookX, (double)lookZ));
+        double newLookAngle = Math.toDegrees(Math.atan2(lookX, lookZ));
         double newNewAngle = this.loopAngle(newAngle - newLookAngle) + (double)fov;
         newNewAngle = this.loopAngle(newNewAngle);
         if (newNewAngle > 0.0D && newNewAngle < (double)(fov * 2.0F)) {
@@ -83,11 +83,11 @@ public class DwellerStareGoal extends Goal {
         float yFov = fov * yFovMod;
         Vec2 yDist = new Vec2((float)Math.sqrt((b.x - a.x) * (b.x - a.x) + (b.z - a.z) * (b.z - a.z)), (float)(b.y - a.y));
         yDist = yDist.normalized();
-        double yAngle = Math.toDegrees(Math.atan2((double)yDist.x, (double)yDist.y));
+        double yAngle = Math.toDegrees(Math.atan2(yDist.x, yDist.y));
         float lookY = (float)pendingTarget.getViewVector(1.0F).y;
-        Vec2 lookDist = new Vec2((float)Math.sqrt((double)(lookX * lookX + lookZ * lookZ)), lookY);
+        Vec2 lookDist = new Vec2((float)Math.sqrt((lookX * lookX + lookZ * lookZ)), lookY);
         lookDist = lookDist.normalized();
-        double yLookAngle = Math.toDegrees(Math.atan2((double)lookDist.x, (double)lookDist.y));
+        double yLookAngle = Math.toDegrees(Math.atan2(lookDist.x, lookDist.y));
         double newYAngle = this.loopAngle(yAngle - yLookAngle) + (double)yFov;
         newYAngle = this.loopAngle(newYAngle);
         if (newYAngle > 0.0D && newYAngle < (double)(yFov * 2.0F)) {
@@ -120,8 +120,8 @@ public class DwellerStareGoal extends Goal {
     public void tick() {
         this.tickStareClock();
         if (this.shouldLeave && (!this.isPlayerLookingTowards() || !this.inPlayerLineOfSight())) {
-            this.cavedweller.playDisappearSound(); // Leaving custom method untouched
-            this.cavedweller.discard(); // Leaving custom method untouched
+            this.cavedweller.playDisappearSound();
+            this.cavedweller.discard();
         }
 
         this.cavedweller.getLookControl().setLookAt(this.cavedweller.getTarget(), 180.0F, 1.0F);
